% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnads-package.R
\docType{package}
\name{dnads-package}
\alias{dnads}
\alias{dnads-package}
\title{dnads: Download, Read and Design PNADs}
\description{
With simple and small functions, the package provide tools to list, download, read and design PNAD data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hhmacedo/dnads}
  \item Report bugs at \url{https://github.com/hhmacedo/dnads/issues}
}

}
\author{
\strong{Maintainer}: Hugo Macedo \email{hhmacedo@gmail.com}

}
\keyword{internal}
