% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnad_design.R
\name{pnad_design}
\alias{pnad_design}
\title{Implement survey design.}
\usage{
pnad_design(df)
}
\arguments{
\item{df}{Dataframe to be read}
}
\value{
A survey object.
}
\description{
Implement survey design.
}
