% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnad_list.R
\name{pnad_list}
\alias{pnad_list}
\title{Lists available years or required files.}
\usage{
pnad_list(year = NULL, files = FALSE)
}
\arguments{
\item{year}{Selected year to return files.}

\item{files}{If TRUE, returns required files for a given year.}
}
\value{
A vector of years or files urls.
}
\description{
Lists available years or required files.
}
\examples{
# Print years available
pnad_list()

# List files for 2015
pnad_list(files = TRUE, year = 2015)
}
