% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnad_read.R
\name{pnad_read}
\alias{pnad_read}
\title{Import the data from PNAD using SAS input}
\usage{
pnad_read(hh_data, hh_input, prs_data, prs_input, psu_strat, vars)
}
\arguments{
\item{hh_data}{Household data file.}

\item{hh_input}{Household input SAS file.}

\item{prs_data}{Person data file.}

\item{prs_input}{Person input SAS file.}

\item{psu_strat}{File containing PSU and Strat.}

\item{vars}{Selected variables.}
}
\value{
A tibble
}
\description{
The pnad_read() function is responsible for read the data using the input
file. The variables needed to use the pnad_design() function will always be
imported automatically, even if not selected.
}
\examples{
\dontrun{
# Import household data
pnad2008_dom <- pnad_read(hh_data = "DOM2008.txt", hh_input = "input DOM2008.txt")

# Import household and person data, but just mandatory vars and V8005.
pnad2008 <- pnad_read(hh_data = "DOM2008.txt", hh_input = "input DOM2008.txt",
                      prs_data = "PES2008.TXT", prs_input = "input PES2008.txt",
                      vars = c("V8005"))
}
}
