% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnad_design.R
\name{pnad_design}
\alias{pnad_design}
\title{Implement a complex survey design}
\usage{
pnad_design(df)
}
\arguments{
\item{df}{Data frame from PNAD}
}
\value{
An object of class survey.design.
}
\description{
Implement a complex survey design
}
