% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnad_get.R
\name{pnad_get}
\alias{pnad_get}
\title{Download PNAD files and read them.}
\usage{
pnad_get(year = NULL, design = TRUE)
}
\arguments{
\item{year}{A year must be provided.}

\item{design}{TRUE if you want a survey object, FALSE if you want a dataframe.}
}
\value{
A survey object or a tibble.
}
\description{
Download PNAD files and read them.
}
\examples{
\dontrun{
pnad_get(year = 2015)
}

}
